+++ src/telas/Home.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import Produtores from './Produtores';
+
+export default function Home() {
+  return <Produtores melhoresProdutores={false} />;
+}
