+++ src/telas/MelhoresProdutores.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import Produtores from './Produtores';
+
+export default function MelhoresProdutores() {
+  return <Produtores melhoresProdutores />;
+}
